lightclient: Refactor modules
