+++ lightclient/src/platform/wasm_helpers.rs
